% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheet.R
\name{sheet}
\alias{sheet}
\title{Function to create sheets}
\usage{
sheet(
  ...,
  row.names = NULL,
  check.rows = FALSE,
  check.names = TRUE,
  fix.empty.names = TRUE,
  stringsAsFactors = FALSE
)
}
\description{
Function to create sheets
}
